+++ CHANGELOG.md
@@ -6,7 +6,8 @@
     - button: add `decoupled` property (restrict visual feedback to values received from osc/midi messages)
     - canvas: iOS: trigger `onTouch` when `event.force`, `event.altitudeAngle` or `event.azimuthAngle` updates
     - knob: add `snap-alt` mode (snap mode that allows jumping between range.min and range.max)
-
+    - all: add `borderRadius` property
+    
 - scripting
     - add `updateCanvas()` function (forces a canvas widget to redraw)